Enable using an installed chroot as a prefix
The idea is you can install i.e. a different PHP version in that
chroot, which is usable even without the chroot, and then set -P
plus -c when making a new site.

Caveat: it needs to be set in php.ini, so the INI template needs
changed and -I must be used.