HotBackup Off in http template
Default over the years changed to HotBackup On, causing the launch of an unexpected second set of PHP jobs. The only purpose of the hot backup jobs is if the primary job somehow ends unexpectedly, or there is an "internal restart" (rare) then the restart would be faster. I believe this is an edge case that is not worth the cost, confusion, and management burden of the additional jobs.